RAG/requirements.txt
fastapi
uvicorn
pydantic
pymilvus
sentence-transformers
ollama
python-multipart
pandas
requests
python-dotenv 